Tweak DISPLAY_DEVICE based on macOS release
Do not try and coerce the screen resolution on macOS. Resolution is defined via System Preferences only on macOS. 

qxl is the display device on Catalina and Big Sur and VGA is used on Mahave and High Sierra.

bochs-display also appears to work correctly on Big Sur providing the QEMU moniotor is set to none, which Quickemu does by default.