Disable all network interfaces during Windows install. Close #115
During the Windows install network interfaces are completely disabled! This is to allow those users who want to create local accounts to do so.

A side affect is that the first time Windows boots with the QEMU network enabled it will blue screen, 
automatically restart and then boot normally with fully functioning networking.,