Revert localiso feature
Mixing an option parser with the existing argv/arvc style parser creates a confusing set of conditionals with large if/then inside a case statement.

On reflection, I don't think this feature has a place in quickget. For anyone with existing iso image, quickemu configurations can be hand crafted. Perhaps the additional of a "quickconf" would be a better way to quickly create VM configs from existing ISO images.