README.md: Update macOS install steps with correct hard disk naming
In a previous commit we dropped usage of VirtIO Block Media from backing the hard disk device in macOS guests. The instructions therefore are updated here to drop the split between macOS versions that we had previously enabled VirtIO Block Media disks and those we didn't. This simplifies the instructions to a single hard disk name for all supported macOS version installations.