Regenerated docs to adopt changes from PR 1140
man pages so far retain much of the installation and usage detail.
The README.md is now not really requiring regeneration as all the details
are referenced in the wiki.  It will be added in a seperate commit which can be ignored
but serves to show the 'corrections' and style tweaks imposed by pandoc.  These principally are line length and whitespace,
 with some markdown pedantry