add extra link filter safeguards
There are  only two  links visible on a web scrape.   The others with similar names are embedded in .js menus.

Add a filter for these other similarities in case, in the future, the other links become visible.

Other filters like ensuring there are digits in the file name could cause future breakage if the format or format sequence became changed.