refactor:  make releases grep line more robust
Header files are not always consistent in their format.
'Location' can sometimes be capitalized, hence the '-i'

Regex can too easily turn into lines of indecipherable hieroglyphics  but in this case, due to variance as cited above, regex would probably be more robust than grepping for 'location:' (with a colon)  and if the header was reorganised then '-m 1' could end up up failing.