refactor:  further improve mx grep
On further research and reflection to 503e083

There are two types of location headers 'Location:' and 'Content-Location:'   so using the regex caret anchor ^'location'   could fail when direct links are used as it would only look for 'location' at the start of the line.

See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers

" An HTTP header consists of its case-insensitive name followed by a colon (:), then by its value."

There is also the issue of different versions of grep having different regex capabilities to consider.

As such, I am now viewing a grep with  -i  'location:'  (with a colon)  as the more reliable general solution and suggest that this is the one to go with.